+++ setup.py
@@ -7,8 +7,6 @@ setup(
     author="ClimateMind",
     url="https://github.com/ClimateMind/climatemind-backend",
     py_modules=[
-        "process_new_ontology_and_visualize",
-        "process_new_ontology_file",
         "config",
     ],
     packages=find_packages(
